outbreak-dashboard/ingest/requirements.txt
feedparser==6.0.11
httpx==0.28.1
PyYAML==6.0.2
